Add files via upload, codes differnt
Added files upload and the codes are a different submission than the one I did with my team. I could not find the file. I tried to download it again from discord, but the file link expired. The pptx is there in a pdf modified to remove faces for privacy. option 3 is my rover, but missing the CAD file I lost. See the READ.ME file